+++ !KalielsTracker/Modules/Options.lua
@@ -4,11 +4,12 @@
 ---
 --- This file is part of addon Kaliel's Tracker.
 
+---@type KT
 local addonName, KT = ...
 
+---@class Options
 local M = KT:NewModule("Options")
 KT.Options = M
-KT.forcedUpdate = false
 
 local ACD = LibStub("MSA-AceConfigDialog-3.0")
 local ACR = LibStub("AceConfigRegistry-3.0")
@@ -33,6 +34,7 @@ local strata = { "LOW", "MEDIUM", "HIGH" }
 local flags = { [""] = "None", ["OUTLINE"] = "Outline", ["OUTLINE, MONOCHROME"] = "Outline Monochrome" }
 local textures = { "None", "Default (Blizzard)", "One line", "Two lines" }
 local modifiers = { [""] = "None", ["ALT"] = "Alt", ["CTRL"] = "Ctrl", ["ALT-CTRL"] = "Alt + Ctrl" }
+local ICON_HEART = "|T"..KT.MEDIA_PATH.."Help\\help_patreon:14:14:0:0:256:32:174:190:0:16|t"
 
 local cTitle = " "..NORMAL_FONT_COLOR_CODE
 local cBold = "|cff00ffe3"
@@ -41,7 +43,7 @@ local beta = "|cffff7fff[Beta]|r"
 local warning = cWarning.."Warning:|r UI will be re-loaded!"
 
 local KTF = KT.frame
-local OTF = ObjectiveTrackerFrame
+local OTF = KT_ObjectiveTrackerFrame
 
 local overlay
 local overlayShown = false
@@ -115,9 +117,12 @@ local defaults = {
 		tooltipShowRewards = true,
 		tooltipShowID = true,
 
+		pettrackerHeaderTitleAppend = true,
+
 		modulesOrder = KT.BLIZZARD_MODULES,
 
 		addonQuestie = false,
+		addonPetTracker = false,
 	},
 	char = {
 		collapsed = false,
@@ -157,7 +162,7 @@ local options = {
 							order = 0.12,
 						},
 						slashCmd = {
-							name = cBold.." /kt|r  |cff808080..............|r  Toggle (expand/collapse) the tracker\n"..
+							name = cBold.." /kt|r  |cff808080...............|r  Toggle (expand/collapse) the tracker\n"..
 									cBold.." /kt config|r  |cff808080...|r  Show this config window\n",
 							type = "description",
 							width = "double",
@@ -185,14 +190,15 @@ local options = {
 							end,
 							order = 0.4,
 						},
-						supportersSpacer = {
-							name = " ",
+						about = {
+							name = " Made with "..ICON_HEART.." since 2012\n"..
+									" |cff999999by "..KT.AUTHOR,
 							type = "description",
 							width = "normal",
 							order = 0.51,
 						},
 						supportersLabel = {
-							name = "                |cff00ff00Become a Patron",
+							name = "                   |cff00ff00Become a Patron",
 							type = "description",
 							width = "normal",
 							fontSize = "medium",
@@ -477,9 +483,7 @@ local options = {
 							values = WidgetLists.statusbar,
 							set = function(_, value)
 								db.progressBar = value
-								KT.forcedUpdate = true
-								ObjectiveTracker_Update()
-								KT.forcedUpdate = false
+								KT:SendSignal("OPTIONS_CHANGED", true)
 							end,
 							order = 2.9,
 						},
@@ -498,10 +502,8 @@ local options = {
 							values = WidgetLists.font,
 							set = function(_, value)
 								db.font = value
-								KT.forcedUpdate = true
-								KT:SetText()
-								ObjectiveTracker_Update()
-								KT.forcedUpdate = false
+								KT:SetText(true)
+								KT:SendSignal("OPTIONS_CHANGED")
 							end,
 							order = 3.1,
 						},
@@ -513,10 +515,8 @@ local options = {
 							step = 1,
 							set = function(_, value)
 								db.fontSize = value
-								KT.forcedUpdate = true
-								KT:SetText()
-								ObjectiveTracker_Update()
-								KT.forcedUpdate = false
+								KT:SetText(true)
+								KT:SendSignal("OPTIONS_CHANGED")
 							end,
 							order = 3.2,
 						},
@@ -533,10 +533,8 @@ local options = {
 							end,
 							set = function(_, value)
 								db.fontFlag = value
-								KT.forcedUpdate = true
-								KT:SetText()
-								ObjectiveTracker_Update()
-								KT.forcedUpdate = false
+								KT:SetText(true)
+								KT:SendSignal("OPTIONS_CHANGED")
 							end,
 							order = 3.3,
 						},
@@ -571,10 +569,8 @@ local options = {
 							type = "toggle",
 							set = function()
 								db.textWordWrap = not db.textWordWrap
-								KT.forcedUpdate = true
-								ObjectiveTracker_Update()
-								ObjectiveTracker_Update()
-								KT.forcedUpdate = false
+								KT:SendSignal("OPTIONS_CHANGED", true)
+								KT:SendSignal("OPTIONS_CHANGED")
 							end,
 							order = 3.6,
 						},
@@ -1130,6 +1126,28 @@ local options = {
 						},
 					},
 				},
+				sec4 = {
+					name = "Addon - PetTracker",
+					type = "group",
+					inline = true,
+					order = 4,
+					args = {
+						pettrackerHeaderTitleAppend = {
+							name = "Show number of owned Pets",
+							desc = "Show number of owned Pets inside the PetTracker header.",
+							type = "toggle",
+							width = "normal+half",
+							disabled = function()
+								return not KT.AddonPetTracker.isLoaded
+							end,
+							set = function()
+								db.pettrackerHeaderTitleAppend = not db.pettrackerHeaderTitleAppend
+								KT.AddonPetTracker:SetPetsHeaderText(true)
+							end,
+							order = 4.1,
+						},
+					},
+				},
 			},
 		},
 		modules = {
@@ -1179,12 +1197,37 @@ local options = {
 							order = 1.11,
 						},
 						addonQuestieDesc = {
-							name = "Questie support adds:\n"..
-									"- context options "..cBold.."Show on Map|r and "..cBold.."Set TomTom Waypoint|r,\n"..
-									"- "..cBold.."Quest Item buttons|r for quests with usable items.",
+							name = "Enables "..cBold.."Show on Map|r and "..cBold.."Set TomTom Waypoint|r options in the "..
+									"Quest context menu, and adds Quest Item buttons for quests with usable items.",
 							type = "description",
 							width = "double",
 							order = 1.12,
+						},
+						addonPetTracker = {
+							name = "PetTracker",
+							desc = "Version: %s",
+							descStyle = "inline",
+							type = "toggle",
+							width = 1.05,
+							confirm = true,
+							confirmText = warning,
+							disabled = function()
+								return not C_AddOns.IsAddOnLoaded("PetTracker")
+							end,
+							set = function()
+								db.addonPetTracker = not db.addonPetTracker
+								if PetTracker.sets then
+									PetTracker.sets.zoneTracker = db.addonPetTracker
+								end
+								ReloadUI()
+							end,
+							order = 1.21,
+						},
+						addonPetTrackerDesc = {
+							name = "Enables display of zone pet tracking inside the tracker and fixes some visual issues.",
+							type = "description",
+							width = "double",
+							order = 1.22,
 						},
 					},
 				},
@@ -1284,7 +1327,7 @@ function GetModulesOptionsTable()
 			order = 0.1,
 		},
 		descDefOrder = {
-			name = "|T:1:42|t"..cTitle.."Default Order",
+			name = "|T:1:20|t"..cTitle.."Default Order",
 			type = "description",
 			width = "normal",
 			fontSize = "medium",
@@ -1293,7 +1336,7 @@ function GetModulesOptionsTable()
 	}
 	if WOW_PROJECT_ID > WOW_PROJECT_CLASSIC then
 		args.descModules = {
-			name = "\n * "..TRACKER_HEADER_SCENARIO.." / |cffff0000"..TRACKER_HEADER_PROVINGGROUNDS.." (not supported)\n",
+			name = "\n * "..TRACKER_HEADER_SCENARIO.." / "..CHALLENGE_MODE.." / |cffff0000"..TRACKER_HEADER_PROVINGGROUNDS.." (not supported)\n",
 			type = "description",
 			order = 20,
 		}
@@ -1305,7 +1348,7 @@ function GetModulesOptionsTable()
 			text = text.." *"
 		end
 
-		defaultModule = OTF.MODULES_UI_ORDER[i]
+		defaultModule = OTF.MODULES[i]
 		defaultText = defaultModule.Header.Text:GetText()
 		if defaultModule == SCENARIO_CONTENT_TRACKER_MODULE then
 			defaultText = defaultText.." *"
@@ -1339,7 +1382,7 @@ function GetModulesOptionsTable()
 			order = i + 0.2,
 		}
 		args["pos"..i.."default"] = {
-			name = "|T:1:55|t|cff808080"..defaultText,
+			name = "|T:1:24|t|cff808080"..defaultText,
 			type = "description",
 			width = "normal",
 			fontSize = "medium",
@@ -1435,6 +1478,11 @@ local function Setup()
 				for i = 1, #db.filterAuto do
 					db.filterAuto[i] = nil
 				end
+				if KT.AddonPetTracker.isLoaded then
+					if PetTracker.sets.zoneTracker then
+						PetTracker.ToggleOption("zoneTracker")
+					end
+				end
 				KT:SetBackground()
 				KT.QuestsCache_Rebuild(true, true)
 				KT.stopUpdate = false
@@ -1474,6 +1522,10 @@ local function Setup()
 		db.messageAchievement = false
 		content.sec2 = nil
 		db.achievementsHeaderTitleAppend = false
+		content.sec4 = nil
+		db.pettrackerHeaderTitleAppend = false
+		addons.sec1.args.addonPetTracker = nil
+		addons.sec1.args.addonPetTrackerDesc = nil
 	end
 end