+++ !KalielsTracker/CHANGELOG.md
@@ -1,4 +1,18 @@
 # Kaliel's Tracker Classic changelog
+### 5.1.3
+- FIXED - occasional error when cold starting WoW
+### 5.1.2
+- FIXED - &quot;Expand after track&quot; option not working for all quests
+### 5.1.1
+- FIXED (Era) - error when tracker is empty (no PetTracker module)
+### 5.1.0
+- ADDED (MoP) - dungeon Challenge Mode support (untested)
+- ADDED - addon support - PetTracker 11.2
+- CHANGED - addon support - Questie 11.3.1
+- CHANGED - PetTracker - improvement
+- FIXED - Help - error if no quest is being tracked
+- FIXED - PetTracker - sometimes SetWidth &quot;nil value&quot; error
+- PERFORMANCE - optimization of zone filtering code
 ### 5.0.1
 - FIXED (quests) - error when filtering by zone with Auto Quest tracking
 ### 5.0.0